Encapsulate the reflective access to the compiler args in its own type
